% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followingNetwork.R
\name{getADJNetDen}
\alias{getADJNetDen}
\title{getADJNetDen function}
\usage{
getADJNetDen(adjMat)
}
\arguments{
\item{adjMat}{is an adjacency matrix of a dominant-distribution network.}
}
\value{
This function returns a value of network density of of a network for a given adjMat.
}
\description{
getADJNetDen is a support function for calculating a network density of a network.
}
\examples{

# Given an example of adjacency matrix
A<-matrix(FALSE,5,5)
A[2,1]<-TRUE
A[c(3,4),2]<-TRUE

# Get a network density of an adjacency matrix

getADJNetDen(adjMat=A)

}
