% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFuncs.R
\name{plotMultipleTimeSeries}
\alias{plotMultipleTimeSeries}
\title{plotMultipleTimeSeries}
\usage{
plotMultipleTimeSeries(TS, strTitle = "Time Series Plot", TSnames)
}
\arguments{
\item{TS}{is a set of time series where \code{TS[i,t,d]} is a numeric value of \code{i}th time series at time \code{t} and dimension \code{d}.}

\item{strTitle}{is a string of the plot title}

\item{TSnames}{is a list of legend of \code{X,Y} where TSnames[1] is a legend of \code{X} and  TSnames[2] is a legend of \code{Y}.}
}
\value{
This function returns an object of ggplot class.
}
\description{
plotMultipleTimeSeries is a function for visualizing time series
}
\examples{
# Run the function
plotMultipleTimeSeries(TS=mFLICA::TS[1:5,1:60,1])

}
